Removed Java compiler warnings from Projector plugin - one left
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@12280 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ plugins/Projector/src/org/micromanager/projector/Galvo.java
@@ -86,7 +86,7 @@ public class Galvo implements ProjectionDevice {
 
             public Double call() {
                try {
-                  return (double) mmc_.getGalvoXRange(galvo_);
+                  return mmc_.getGalvoXRange(galvo_);
                } catch (Exception ex) {
                   return 0.0;
                }
@@ -108,7 +108,7 @@ public class Galvo implements ProjectionDevice {
 
             public Double call() {
                try {
-                  return (double) mmc_.getGalvoYRange(galvo_);
+                  return mmc_.getGalvoYRange(galvo_);
                } catch (Exception ex) {
                   return 0.0;
                }
